requirements.txt

streamlit
openai
faiss-cpu
sentence-transformers
PyPDF2
python-dotenv
langchain